Changes to Support Web Tool
Assorted changes to support a Web based version of the tool that runs in a container:
Ability to specify static content that will be served by NGINX and can be retrieved by a browser.
Added examples of static content to project1 example.
Now enforce limitations on folder names (<=30 chars, lowercase, numbers, & underscores only)
Changes port ranges to not include 5000, since this is special to Docker.
Added flag in project file and support for running tool locally (command prompt) or in a container.
Added capabilities to read context in container and also from attached bind mounted directory.
Added exceptions.