First Working Web Tool
Replaced the FontAwesome trash can, eliminating FontAwesome and hundreds of files.
Completed Build Project in the Web tool.
Fixed an issue where select lists and associated buttons were not disabled appropriately.
Fixed an issue where the listing of function arguments and checkbox states were not reset properly.
Fixed an issue where folders were not being referenced correctly in the generated service modules, resulting in the services not working.
Fixed an issue where function names were not appearing in the service modules, resulting in the services not working.
Added instructions for running the Web Tool on port 8080 to avoid conflicts with the services created.
Completed the basic Web Tool, including project builds that can be manually built into working containers/services!